refactoring bloco-11
